[docs] Break up blog template into smaller sections
