+++ docs/src/pages/getting-started/templates/blog/blog-post.1.md
@@ -1,4 +1,4 @@
-## Sample blog post
+# Sample blog post
 
 #### April 1, 2020 by [Olivier](/)
 
@@ -33,14 +33,14 @@ Aenean lacinia bibendum nulla sed consectetur. Etiam porta sem malesuada magna m
 Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum nibh, ut fermentum massa justo
 sit amet risus.
 
- - Praesent commodo cursus magna, vel scelerisque nisl consectetur et.
- - Donec id elit non mi porta gravida at eget metus.
- - Nulla vitae elit libero, a pharetra augue.
+- Praesent commodo cursus magna, vel scelerisque nisl consectetur et.
+- Donec id elit non mi porta gravida at eget metus.
+- Nulla vitae elit libero, a pharetra augue.
 
 Donec ullamcorper nulla non metus auctor fringilla. Nulla vitae elit libero, a pharetra augue.
 
- 1. Vestibulum id ligula porta felis euismod semper.
- 2. Cum sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus.
- 3. Maecenas sed diam eget risus varius blandit sit amet non magna.
+1.  Vestibulum id ligula porta felis euismod semper.
+2.  Cum sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus.
+3.  Maecenas sed diam eget risus varius blandit sit amet non magna.
 
 Cras mattis consectetur purus sit amet fermentum. Sed posuere consectetur est at lobortis.
