+++ docs/src/pages/getting-started/templates/blog/blog-post.2.md
@@ -1,4 +1,4 @@
-## Another blog post
+# Another blog post
 
 #### March 23, 2020 by [Matt](/)
