+++ docs/src/pages/getting-started/templates/blog/blog-post.3.md
@@ -1,4 +1,4 @@
-## New feature
+# New feature
 
 #### March 14, 2020 by [Tom](/)
 
@@ -7,9 +7,9 @@ Aenean lacinia bibendum nulla sed consectetur. Etiam porta sem malesuada magna m
 Fusce dapibus, tellus ac cursus commodo, tortor mauris condimentum nibh,
 ut fermentum massa justo sit amet risus.
 
- - Praesent commodo cursus magna, vel scelerisque nisl consectetur et.
- - Donec id elit non mi porta gravida at eget metus.
- - Nulla vitae elit libero, a pharetra augue.
+- Praesent commodo cursus magna, vel scelerisque nisl consectetur et.
+- Donec id elit non mi porta gravida at eget metus.
+- Nulla vitae elit libero, a pharetra augue.
 
 Etiam porta sem malesuada magna mollis euismod. Cras mattis consectetur purus sit amet fermentum.
 Aenean lacinia bibendum nulla sed consectetur.
